requirements.txt
# Core ML
xgboost>=2.0.0
lightgbm>=4.3.0
scikit-learn>=1.4.0
optuna>=3.6.0

# Data
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# HTTP / Scraping
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Data storage
pyarrow>=15.0.0
fastparquet>=2024.2.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0

# Config & CLI
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0

# Web app
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Calibration & stats
imbalanced-learn>=0.12.0
statsmodels>=0.14.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.29.0
